primerlab::functions
Function fib

pub fn fib(n: usize) -> usize

Returns the nth fibonacci number.

We consider the 0th fibonacci number to be 0, and the first to be 1.

Example

use primerlab::functions::fib;

assert_eq!(fib(2), 1);
assert_eq!(fib(4), 3);
assert_eq!(fib(7), 13);

Note

There are a few ways to implement this. However, one obvious way might time out on Gradescope…

The easiest way to do this is to either use iteration instead of recursion, or to use some sort of helper function (which you can define somewhere else in this file).

Feel free to look this up online if you’re having trouble!