primerlab::functions
Function nth_prime

pub fn nth_prime(n: usize) -> usize

Given a number n, return the nth prime. Refer to the test cases below for more details.

For example, the 0th prime is 2, and the 1st prime is 3, then the 2nd prime is 5, etc.

Example

use primerlab::functions::nth_prime;

let n = 4;
assert_eq!(nth_prime(n), 11);

let n = 20;
assert_eq!(nth_prime(n), 73);

Suggestions

You can look here for more primes, but note that Wikipedia 1-indexes them.